quickstarts/discord-bot-cogs/requirements.txt
discord.py
python-dotenv
web3
requests
aiohttp
